MorphologyMetricsOutput moved to scientific/morphology_metrics. MorphologyMetricsExample.. replaces old MorphologyMetrics which is removed
