+++ obi_one/__init__.py
@@ -30,8 +30,7 @@ from obi_one.scientific.connectivity_matrix_extraction.connectivity_matrix_extra
 from obi_one.scientific.basic_connectivity_plots.basic_connectivity_plots import BasicConnectivityPlots, BasicConnectivityPlot
 from obi_one.scientific.folder_compression.folder_compression import FolderCompressions, FolderCompression
 from obi_one.scientific.morphology_containerization.morphology_containerization import MorphologyContainerizationsForm, MorphologyContainerization
-from obi_one.scientific.morphology_metrics.morphology_metrics import MorphologyMetricsForm, MorphologyMetrics
-from obi_one.scientific.morphology_metrics.morphology_metrics_example import MorphologyMetricsExampleForm, MorphologyMetricsExample
+from obi_one.scientific.morphology_metrics.morphology_metrics import MorphologyMetricsForm, MorphologyMetrics, MorphologyMetricsOutput
 
 from obi_one.scientific.test_forms.test_form_single_block import SingleBlockGenerateTestForm, SingleBlockGenerateTest, SingleBlockEntitySDKTestForm, SingleBlockEntitySDKTest, MultiBlockEntitySDKTestForm, MultiBlockEntitySDKTest
