--- obi_one/scientific/morphology_metrics/morphology_metrics_example.py
@@ -1,78 +1,0 @@
-from typing import Union
-from typing import ClassVar
-
-from obi_one.core.form import Form
-from obi_one.core.block import Block
-from obi_one.core.single import SingleCoordinateMixin
-from obi_one.core.path import NamedPath
-from obi_one.core.base import OBIBaseModel
-from obi_one.database.db_classes import *
-
-
-
-
-class MorphologyMetricsExampleForm(Form):
-    """
-    """
-
-    single_coord_class_name: ClassVar[str] = "MorphologyMetricsExample"
-    name: ClassVar[str] = "1. Morphology metrics example name"
-    description: ClassVar[str] = "2. Morphology metrics example description"
-
-    class Initialize(Block):
-        morphology: ReconstructionMorphology | ReconstructionMorphologyFromID | list[Union[ReconstructionMorphology, ReconstructionMorphologyFromID]] = Field(description="3. Morphology description")
-
-    initialize: Initialize
-
-    def save(self, circuit_entities):
-        """
-        Add entitysdk calls to save the collection
-        """
-        pass
-
-
-class MorphologyMetricsExampleRunOutput(OBIBaseModel):
-    features: dict = Field(
-    default_factory=dict,
-    description="4. Morphology metrics example output",
-)
-
-
-
-import traceback
-import neurom
-import numpy as np
-class MorphologyMetricsExample(MorphologyMetricsExampleForm, SingleCoordinateMixin):
-    """
-    Code orgininally created for https://github.com/openbraininstitute/neuroagent by:
-    Boris Bergsma, Nicolas Frank, Jan Krepl
-    in https://github.com/openbraininstitute/neuroagent/blob/main/backend/src/neuroagent/tools/morphology_features_tool.py
-    """
-    _features: dict = {}
-
-
-    def run(self):
-        
-        try:
-            self._features = {
-                "soma_radius [µm]": neurom.get("soma_radius", self.initialize.morphology.neurom_morphology),
-                "soma_surface_area [µm^2]": neurom.get("soma_surface_area", self.initialize.morphology.neurom_morphology),
-            }
-
-        except Exception as e:
-            traceback.print_exception(e)
-            
-        
-
-    def save(self):
-        """
-        Add entitysdk calls to save the single instance
-        """
-        return None        
-
-    def data(self) -> MorphologyMetricsExampleRunOutput:
-        """
-        Return the data to the client
-        """
-        return MorphologyMetricsExampleRunOutput(features=self._features)
-
